src/edu-api/pyproject.toml
[project]
name = "edu-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edu-core",
    "edu-ai",
    "edu-queue",
    "fastapi>=0.124.4",
    "langchain-openai>=1.1.2",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "scalar-fastapi>=1.0.0",
    "uvicorn>=0.38.0",
]

[tool.uv.sources]
edu-core = { workspace = true }
edu-queue = { workspace = true }
edu-ai = { workspace = true }
